another edit on crud in Brand And Category And Product

+++ routes/admin.php
@@ -51,7 +51,7 @@ Route::group([
 });
 
 Route::controller(ProductController::class)->group(function () {
-    Route::get('products','index')->middleware('auth:admin')->name('admin.products');
+    Route::get('products','index')->middleware('auth:admin')->name('admin.products.index');
     Route::post('products/store','store')->middleware('auth:admin')->name('admin.products.store');
     Route::get('products/create','create')->middleware('auth:admin')->name('admin.products.create');
     Route::get('products/edit/{products}','edit')->middleware('auth:admin')->name('admin.products.edit');
@@ -59,12 +59,13 @@ Route::controller(ProductController::class)->group(function () {
     Route::get('products/delete','delete')->middleware('auth:admin')->name('admin.products.delete');
 });
 Route::controller(BrandController::class)->group(function () {
-    Route::get('brands','index')->name('admin.brands');
-    Route::post('brands/store','store')->name('admin.brands.store');
-    Route::get('brands/create','create')->name('admin.brands.create');
-    Route::get('brands/edit/{brands}','edit')->name('admin.brands.edit');
-    Route::put('brands/update/{id}','update')->name('admin.brands.update');
-    Route::delete('brands/delete/{id}','delete')->name('admin.brands.delete');
+    Route::get('brands','index')->middleware('auth:admin')->name('admin.brands.index');
+    Route::post('brands/store','store')->middleware('auth:admin')->name('admin.brands.store');
+    Route::get('brands/create','create')->middleware('auth:admin')->name('admin.brands.create');
+    Route::get('brands/show/{id}','show')->middleware('auth:admin')->name('admin.brands.show');
+    Route::get('brands/edit/{brands}','edit')->middleware('auth:admin')->name('admin.brands.edit');
+    Route::put('brands/update/{id}','update')->middleware('auth:admin')->name('admin.brands.update');
+    Route::delete('brands/delete/{id}','delete')->middleware('auth:admin')->name('admin.brands.delete');
 });
 
 // Route::resource('/products' , ProductController::class);
@@ -72,8 +73,8 @@ Route::controller(BrandController::class)->group(function () {
 // Route::resource('/carts' , CartController::class);
 // Route::get('/cartitems{id}/add' , [CartitemController::class , 'add'])->name('carts.add');
 // Route::get('/products{id}/delete' , [ProductController::class , 'delete'])->name('products.delete');
-Route::get('/brand{id}/delete' , [BrandController::class , 'delete'])->name('brand.delete');
-Route::get('/category{id}/delete' , [CategoryController::class , 'delete'])->name('category.delete');
+// Route::get('/brand{id}/delete' , [BrandController::class , 'delete'])->name('brand.delete');
+// Route::get('/category{id}/delete' , [CategoryController::class , 'delete'])->name('category.delete');
 // Route::get('/cartitems{id}/delete' , [CartitemController::class , 'delete'])->name('cartitems.delete');
 // Route::get('/cart{id}/creates' , [CartController::class , 'create_cart'])->name('cart.creates');
